[CHE-445] Write Example project from scratch using storyboard.
