+++ Example/PrimerSDK/AppDelegate.swift
@@ -14,15 +14,15 @@ class AppDelegate: UIResponder, UIApplicationDelegate {
     var window: UIWindow?
 
     func application(_ application: UIApplication, didFinishLaunchingWithOptions launchOptions: [UIApplicationLaunchOptionsKey: Any]?) -> Bool {
-        let nc = UINavigationController()
-        nc.navigationBar.barTintColor = UIColor(red: 240/255, green: 97/255, blue: 91/255, alpha: 1)
-        nc.navigationBar.tintColor = .white
-        nc.navigationBar.titleTextAttributes = [.foregroundColor: UIColor.white]
-        let vc = ViewController()
-        nc.viewControllers = [vc]
-        window?.rootViewController = nc
-        window?.makeKeyAndVisible()
+        customizeAppearance()
         return true
+    }
+    
+    private func customizeAppearance() {
+        UINavigationBar.appearance().barTintColor = UIColor(red: 240/255, green: 97/255, blue: 91/255, alpha: 1)
+        UINavigationBar.appearance().titleTextAttributes = [.foregroundColor: UIColor.white]
+        UIBarButtonItem.appearance().tintColor = .white
+        UINavigationBar.appearance().tintColor = .white
     }
 
 }
