+++ Example/PrimerSDK/AppViewController.swift
@@ -1,0 +1,13 @@
+//
+//  AppViewController.swift
+//  PrimerSDK_Example
+//
+//  Created by Evangelos Pittas on 12/4/21.
+//  Copyright © 2021 CocoaPods. All rights reserved.
+//
+
+import UIKit
+
+class AppViewController: UIViewController {
+    
+}
