+++ Example/PrimerSDK/MerchantCheckoutViewController+Helpers.swift
@@ -1,0 +1,35 @@
+//
+//  MerchantCheckoutViewController+Helpers.swift
+//  PrimerSDK_Example
+//
+//  Created by Evangelos Pittas on 12/4/21.
+//  Copyright © 2021 CocoaPods. All rights reserved.
+//
+
+import UIKit
+
+extension MerchantCheckoutViewController {
+    
+    // MARK: - GENERAL HELPERS
+    
+    internal func showResult(error: Bool) {
+        let title = error ? "Error!" : "Success!"
+        let alert = UIAlertController(title: title, message: nil, preferredStyle: .alert)
+        alert.addAction(UIAlertAction(title: "Close", style: .cancel, handler: nil))
+        self.present(alert, animated: true)
+    }
+    
+    internal func addSpinner(_ child: SpinnerViewController) {
+        addChildViewController(child)
+        child.view.frame = view.frame
+        view.addSubview(child.view)
+        child.didMove(toParentViewController: self)
+    }
+    
+    internal func removeSpinner(_ child: SpinnerViewController) {
+        child.willMove(toParentViewController: nil)
+        child.view.removeFromSuperview()
+        child.removeFromParentViewController()
+    }
+    
+}
