+++ Example/PrimerSDK/UIViewController+API.swift
@@ -1,0 +1,60 @@
+//
+//  UIViewController+API.swift
+//  PrimerSDK_Example
+//
+//  Created by Evangelos Pittas on 12/4/21.
+//  Copyright © 2021 CocoaPods. All rights reserved.
+//
+
+import UIKit
+
+// MARK: - API HELPER
+
+struct AuthorizationRequest: Encodable {
+    let paymentMethod: String
+    let amount: Int
+    let type: String?
+    var capture: Bool
+    let currencyCode: String
+}
+
+enum NetworkError: Error {
+    case missingParams
+    case unauthorised
+    case timeout
+    case serverError
+    case invalidResponse
+    case serializationError
+}
+
+extension UIViewController {
+
+    func callApi(_ req: URLRequest, completion: @escaping (_ result: Result<Data, Error>) -> Void) {
+        URLSession.shared.dataTask(with: req, completionHandler: { (data, response, err) in
+
+            if err != nil {
+                completion(.failure(NetworkError.serverError))
+                return
+            }
+
+            guard let httpResponse = response as? HTTPURLResponse else {
+                completion(.failure(NetworkError.invalidResponse))
+                return
+            }
+
+            if (httpResponse.statusCode < 200 || httpResponse.statusCode > 399) {
+                completion(.failure(NetworkError.invalidResponse))
+                return
+            }
+
+            guard let data = data else {
+                completion(.failure(NetworkError.invalidResponse))
+                return
+            }
+
+            completion(.success(data))
+
+        }).resume()
+    }
+
+}
